requirements.txt
Pillow
pyzbar
opencv-python
numpy
reportlab
pytest
